Create Day1:Arithmetic operators
In this challenge, we practice using arithmetic operators. Check out the attached tutorial for resources.
Task
Complete the following functions in the editor below:

getArea(length, width): Calculate and return the area of a rectangle having sides length and width.
getPerimeter(length, width): Calculate and return the perimeter of a rectangle having sides  length and width .
The values returned by these functions are printed to stdout by locked stub code in the editor.